% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{addAnalysis}
\alias{addAnalysis}
\title{Adds a new analysis to a jamovi module}
\usage{
addAnalysis(name, path = ".")
}
\arguments{
\item{name}{the name for the new analysis}
}
\description{
Adds a new analysis to a jamovi module
}

