% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{check}
\alias{check}
\title{Check that jmvtools is able to find jamovi}
\usage{
check(home = NULL)
}
\arguments{
\item{home}{path to a local jamovi installation}
}
\description{
Check that jmvtools is able to find jamovi
}
