% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{create}
\alias{create}
\title{Create an empty jamovi module}
\usage{
create(path = ".", home = NULL, gitignore = TRUE)
}
\arguments{
\item{path}{location to create the new module (the name of the module is inferred from the path)}
}
\description{
Creates an empty jamovi module. Astute observers will notice that empty
jamovi modules are the same as empty R packages.
}
