% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggrotate_xy.R
\name{ggrotate_xy}
\alias{ggrotate_xy}
\title{Rotate a 2D plot by specified number of degrees}
\usage{
ggrotate_xy(df, x, y, deg = 45)
}
\arguments{
\item{df}{Original dataframe object}

\item{x}{Name of column containing data plotted on x-axis}

\item{y}{Name of column containing data plotted on x-axis}

\item{deg}{Degrees to rotate the plotted data}
}
\value{
dataframe with x and y coordinates rotated by specified number of degrees
}
\description{
Rotate a 2D plot by specified number of degrees
}
\note{
Mostly using this to rotate plots from Slide-seq spatial transcriptomics data for figure prep.
}
