% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{i18nCreate}
\alias{i18nCreate}
\title{Create translations}
\usage{
i18nCreate(code, pkg = ".", home = NULL)
}
\arguments{
\item{pkg}{path to a local directory containing the module source}

\item{home}{path to a local jamovi installation}
}
\description{
Create translations
}
