% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{i18nUpdate}
\alias{i18nUpdate}
\title{Update translations}
\usage{
i18nUpdate(code = "*", pkg = ".", home = NULL, debug = FALSE)
}
\arguments{
\item{pkg}{path to a local directory containing the module source}

\item{home}{path to a local jamovi installation}
}
\description{
Update translations
}
