% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{install}
\alias{install}
\title{Build and install a local jamovi module into jamovi}
\usage{
install(pkg = ".", home = NULL)
}
\arguments{
\item{pkg}{path to a local directory containing the module source}

\item{home}{path to a local jamovi installation}
}
\description{
Build and install a local jamovi module into jamovi
}

