% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmv_datatable.R
\name{jmv_datatables}
\alias{jmv_datatables}
\title{Interactive data tables}
\usage{
jmv_datatables(df, digits = 2, ...)
}
\arguments{
\item{df}{Data frame}

\item{digits}{Number of digits to include in formatted numbers.}

\item{...}{Additional parameters to pass to \code{DT::datatable}}
}
\value{
Embedded HTML data table, \code{DT::datatable} object
}
\description{
Interactive data tables
}
\examples{

jmv_datatables(mtcars)

mtcars |> jmv_datatables()

}
