% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmv_plotly_linked.R
\name{jmv_plotly_linked}
\alias{jmv_plotly_linked}
\title{Adding clickable links to data points on ggplot object}
\usage{
jmv_plotly_linked(p)
}
\arguments{
\item{p}{ggplot object, with URL mapped to \code{customdata} inside aes()}
}
\value{
Interactive plot, output by \code{ggplotly()}, with clickable data points that open associated URLs
}
\description{
Adding clickable links to data points on ggplot object
}
\examples{
p <- mtcars |>
    tibble::rownames_to_column(var="vehicle") |>
    dplyr::mutate(link = paste0("https://www.google.com/search?q=", vehicle)) |>
    ggplot2::ggplot(ggplot2::aes(mpg, hp, customdata=link)) + ggplot2::geom_point()
jmv_plotly_linked(p)


}
