% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmv_spatial_umap_annimation.R
\name{jmv_spatial_umap_animation}
\alias{jmv_spatial_umap_animation}
\title{Make animation of data points from spatial to UMAP space}
\usage{
jmv_spatial_umap_animation(seurat_object)
}
\arguments{
\item{seurat_object}{Spatial transcriptomics data in format of Seurat objec that contains UMAP dimensional reduction}
}
\value{
gganimate animation object
}
\description{
Make animation of data points from spatial to UMAP space
}
\examples{
\dontrun{
library(jmvtools)
data <- jmvtools:::spatial_object
jmv_spatial_umap_animation(data)
}

}
