% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmv_superplots.R
\name{jmv_superplots}
\alias{jmv_superplots}
\title{SuperPlots}
\usage{
jmv_superplots(df, groups, replicate, value, font_size = 20, ylab = "")
}
\arguments{
\item{df}{data frame containing data to be plotted}

\item{groups}{name of data frame column containing group levels information}

\item{replicate}{name of data frame column containing replicate levels information}

\item{value}{name of data frame column containing values to be plotted on y-axis for each group, replicate}

\item{font_size}{font size parameter passed to cowplot::theme_cowplot() function}

\item{ylab}{Label for the y-axis of plot}
}
\value{
ggplot object
}
\description{
SuperPlots
}
\examples{
df_1 <- data.frame(value = sample(100:200, 100, replace=TRUE),
                  group = rep(1:4, 25),
                  replicate = 1)
df_2 <- data.frame(value = sample(100:200, 100, replace=TRUE),
                   group = rep(1:4, 25),
                   replicate = 2)

df_3 <- data.frame(value = sample(100:200, 100, replace=TRUE),
                   group = rep(1:4, 25),
                   replicate = 3)

df <- do.call(rbind, list(df_1, df_2, df_3))
jmv_superplots(df, groups = group, replicate = replicate, value = value)
}
