% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/named_group_split.R
\name{named_group_split}
\alias{named_group_split}
\title{Perform a named group split.}
\usage{
named_group_split(.tbl, ...)
}
\arguments{
\item{.tbl}{tibble/dataframe input object}

\item{...}{additional params passed to dplyr::group_by()}
}
\value{
List of data frames, named based on the value of the grouping variable.
}
\description{
Splits a dataframe into a list, with each list object named based on the value of the grouping variable.
}
\note{
stolen from https://github.com/tidyverse/dplyr/issues/4223
}
\examples{

mtcars |> named_group_split(cyl)

}
