% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{prepare}
\alias{prepare}
\title{Prepare a jamovi source module}
\usage{
prepare(pkg = ".")
}
\arguments{
\item{pkg}{path to a local directory containing the module source}
}
\description{
Prepare a jamovi source module
}

