% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{submit}
\alias{submit}
\title{Submit a jamovi module to the jamovi store}
\usage{
submit(pkg = ".")
}
\arguments{
\item{pkg}{path to a local directory containing the module source}
}
\description{
Submit a jamovi module to the jamovi store
}

